+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Enhancements
 
+  * [IEx] Add `alchemist-iex-recompile-this-buffer` which recompiles and reloads the current buffer in the IEx process.
   * [Help] Add `alchemist-help` module prompt. It's loads all available Modules in the current environemnt which have a proper moduledoc.
   * [Help] Add `alchemist-help-history` to `alchemist-help-minor-mode`. `alchemist-help-history` lookup straight for the stored search text, without completion.
   * [Mix] Add mix command prompt with completion