+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## v0.12.0-dev (unreleased)
+## v0.12.0 (2015-01-08)
 
 ### Enhancements